fix the "dock hang" bug.
there's two problems here. the main one is that in one-at-a-time stations, the
docking sequence would only begin if the ship hit the docking surface on stage
0. since (long ago) the addition of the docking approach as stage 1 this meant
the condition would never be satisfied

the flag test masked the issue. it caused the stage test to never be carried
out for port 0

closes #32