Fuel balance: Eagle MK-II
A very slight improvement over the Eagle Long Range Fighter. These two ships
are otherwise very close. The hull mass is reduced for an extra tonne of fuel
tank. Although the total burn time is identical, the reduced mass when empty
gives the MK-II an additional 20,000km/s delta-V on one fill-up. That's
a 10% improvement.