Fuel balance: Turtle
This beast has main thrusters in all six directions, and they're all very
powerful indeed. This ship's thrust force is in a league of its own.

According to the author's notes, it's a high-G, heavily armoured spacecraft.
In lie with this, I've given over more mass to the hull, and less to the
fuel tank. I've also given it very poor efficiency; a ship like this should
be high maintenance, and owners can expect to the kill engines and coast, or
carry additional propellant, even for journeys of normal length.