Output console command before its execution
In the previous code, anything printed within the Lua code would be
printed before the command. Also, the command would not be printed in
case of runtime error, while it is still a valid command, which is
semantically wrong.

Example:

> fonction bogus () print('test') end
test
> bogus() -- the print is executed before outputting the command
>

This patch moves the command printing right before the command
execution, but still after the parsing.