avoid Intel GMA 900 error with 1xN size textures (thanks Skodyn!)
See #1439. Comprehensive testing by Skodyn indicates that the problem is
something to do with mipmapping a very narrow (1 pixel wide) texture.
Most likely this is going wrong during mipmap generation. Eventually we
may want to go back to a 1xN texture and generate mipmaps on the CPU, but
for now using a wider texture should fix the problem.