[valgrind] LmrGeomBuffer: use PushCurOp and clear curOp after pushing
This is to fix a valgrind warning that model cache files are written
with some uninitialised data. Note that this warning does not indicate
any error in the code -- the uninitialised data should never be used,
and if that is the case then it doesn't matter. However, fixing the
warning seems like a good idea in the interests of safety and reducing
the potential for future heisenbugs.