Add Zoom level readout to SectorView + zoom tweaks
Tweaked a lot of the internal scale calculations to make the readout and the
zoom level sync up.
The above causes more systems are drawn per pont on the zoom scale, so the
point at which we stop adding new systems as we zoom out has been reduced
internally is still but bigger in termss of systems drawn than before.
Altered maximum zoom so that there that you can now zoom out a bit beyond the
point where we stop drawing new systems.