Add a Primitive Spatial Index for Factions
Introducing octsapling, it's sort of like an octree (I think!),
but since it's only one level deep you can't really call it a tree.
Lucky for me, as that's about as much as I can get my head around.
It reduces the number of distance checks needed to get a Faction
for a system from the full 103 every time to 20-40 depending on
where the system is.