Skip SectorView::ShrinkCache when it clearly not necessary
SectorView::ShrinkCache was attempting to shrink the cache every
single Update().
I've changed it to only do it's thing (however flawed) if the
block of sectors we should be caching has actually changed.
It's still probably too aggressive, but this brings it down to
being in the noise, when profiling scrolling in Far view.