extract Orbit from StarSystem and display player orbit in SystemView
The Orbit structure from StarSystem.cpp is pulled out into its own
files Orbit.h/cpp, and other orbital calculations are turned into
static functions on that struct.

Orbit struct and calculations are also altered to work for hyperbolic
trajectories.

DynamicBody gains a ReturnOrbit method to compute the orbit of the
body, and this is used to plot the player's orbit in the System View.

Note: This was originally a 15-commit PR from irigi.
I (@johnbartholomew) am squashing it into one commit because the
history is messy but I don't want it to lose irigi as git author.